+++ app/src/main/java/com/example/bronnbakestimer/util/PhoneVibrator.kt
@@ -75,11 +75,11 @@ class PhoneVibrator(private val context: Context) : IPhoneVibrator {
 
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
                 val effect =
-                    VibrationEffect.createOneShot(Constants.HALF_SECOND_MILLIS, VibrationEffect.DEFAULT_AMPLITUDE)
+                    VibrationEffect.createOneShot(Constants.HALF_SECOND_MILLIS.toLong(), VibrationEffect.DEFAULT_AMPLITUDE)
                 vibrator.vibrate(effect)
             } else {
                 @Suppress("DEPRECATION")
-                vibrator.vibrate(Constants.HALF_SECOND_MILLIS)
+                vibrator.vibrate(Constants.HALF_SECOND_MILLIS.toLong())
             }
         }
     }
